Join/signUpTag 폴더구조 정리
