+++ src/pages/Join/SignupTag/SignUpTag.styles.js
@@ -1,0 +1,51 @@
+import styled from 'styled-components';
+
+export const BasketContainer = styled.div`
+  display: flex;
+  flex-direction: row;
+`;
+
+export const BasketWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  margin: auto;
+`;
+export const Container = styled.div`
+  font-family: Arial, sans-serif;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+`;
+export const Wrapper = styled.div`
+  margin: 2rem 0;
+`;
+export const Button = styled.button`
+  margin: 10px 0;
+  width: fit-content;
+`;
+
+export const IntroText = styled.div`
+  margin-top: 10rem;
+  font-size: 24px;
+  font-weight: bold;
+  margin-bottom: 20px;
+`;
+
+export const Basket = styled.div`
+  display: flex;
+  flex-direction: column;
+  background: #d9d9d9;
+  width: 279px;
+  height: 279px;
+  border-radius: 10px;
+  margin: 0 1rem;
+  padding: 1rem;
+  /* min-height: 30px;
+  position: fixed;
+  top: 200px;
+  right: -10px;
+  z-index: 1000; */
+`;
+
+export const SubmitButton = styled.button``;
